src/core/configs.py
from dotenv import load_dotenv
import os

load_dotenv()


class Settings:
    PROJECT_NAME: str = "API de Gestão de Tarefas"
    PROJECT_VERSION: str = "1.0.0"
    SECRET_KEY: str = os.getenv("SECRET_KEY")
    ALGORITHM: str = os.getenv("ALGORITHM")