requirements.txt
beautifulsoup4
pandas
requests
fastapi
fastapi[standard]
uvicorn
sqlalchemy
colorlog
python-jose[cryptography]